月間情報明細Activityを開発

+++ app/src/main/java/com/takassoftware/managey/Model/MonthlyDetailCardUsedListItemModel.kt
@@ -3,5 +3,5 @@ package com.takassoftware.managey.Model
 // 月間情報明細のカード使用リストのアイテム
 data class MonthlyDetailCardUsedListItemModel(
         var cardName : String,
-        var cardUsed : Int
+        var cardUsed : Long
 )
